off_t and ssize_t may be int32 on some platforms and overflow while seeking large file.

+++ include/bdsm/smb_file.h
@@ -118,7 +118,7 @@ ssize_t   smb_fwrite(smb_session *s, smb_fd fd, void *buf, size_t buf_size);
  * @return The current read pointer position or -1 on error
  */
 BDSM_EXPORT
-ssize_t   smb_fseek(smb_session *s, smb_fd fd, off_t offset, int whence);
+int64_t   smb_fseek(smb_session *s, smb_fd fd, int64_t offset, int whence);
 
 /**
  * @brief remove a file on a share.
